# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Notification Service",
    "description": "Multi-channel notification service with priority queue, retry, and idempotency.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/notifications": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Create Notification",
        "operationId": "create_notification_api_v1_notifications_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_NotificationResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get Notification",
        "operationId": "get_notification_api_v1_notifications__notification_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_NotificationDetailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List User Notifications",
        "operationId": "list_user_notifications_api_v1_users__user_id__notifications_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ChannelEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatusEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_NotificationResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/batch": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Create Notification Batch",
        "description": "Bonus: Batch endpoint. Returns 207 Multi-Status with per-item success/failure.",
        "operationId": "create_notification_batch_api_v1_notifications_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/NotificationCreate"
                },
                "type": "array",
                "title": "Notifications In"
              }
            }
          },
          "required": true
        },
        "responses": {
          "207": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_BatchItemResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/users/{user_id}/preferences": {
      "get": {
        "tags": [
          "Preferences"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_api_v1_users__user_id__preferences_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_PreferenceResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Preferences"
        ],
        "summary": "Update Preference",
        "operationId": "update_preference_api_v1_users__user_id__preferences_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PreferenceResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/stats": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Analytics Stats",
        "description": "Bonus: Get analytics grouped by status and channel.",
        "operationId": "get_analytics_stats_api_v1_analytics_stats_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ChannelEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_str__Any__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChannelEnum": {
        "type": "string",
        "enum": [
          "email",
          "sms",
          "push"
        ],
        "title": "ChannelEnum"
      },
      "Envelope_NotificationDetailResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NotificationDetailResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[NotificationDetailResponse]"
      },
      "Envelope_NotificationResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NotificationResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[NotificationResponse]"
      },
      "Envelope_PreferenceResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PreferenceResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[PreferenceResponse]"
      },
...
      "Envelope_dict_str__Any__": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[dict[str, Any]]"
      },
      "Envelope_list_BatchItemResponse__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/BatchItemResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[list[BatchItemResponse]]"
      },
      "Envelope_list_PreferenceResponse__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PreferenceResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[list[PreferenceResponse]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 255,
            "title": "User Id"
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelEnum"
          },
          "priority": {
            "$ref": "#/components/schemas/PriorityEnum",
            "default": "normal"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "template_variables": {
            "additionalProperties": true,
            "type": "object",
            "title": "Template Variables"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "channel"
        ],
        "title": "NotificationCreate"
      },
      "NotificationDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelEnum"
          },
          "priority": {
            "$ref": "#/components/schemas/PriorityEnum"
          },
          "status": {
            "$ref": "#/components/schemas/StatusEnum"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "max_retries": {
            "type": "integer",
            "title": "Max Retries"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/NotificationAttemptSchema"
            },
            "type": "array",
            "title": "Attempts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "channel",
          "priority",
          "status",
          "template_id",
          "payload",
          "idempotency_key",
          "retry_count",
          "max_retries",
          "scheduled_at",
          "created_at",
          "updated_at",
          "attempts"
        ],
        "title": "NotificationDetailResponse"
      },
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelEnum"
          },
          "priority": {
            "$ref": "#/components/schemas/PriorityEnum"
          },
          "status": {
            "$ref": "#/components/schemas/StatusEnum"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "max_retries": {
            "type": "integer",
            "title": "Max Retries"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "channel",
          "priority",
          "status",
          "template_id",
          "payload",
          "idempotency_key",
          "retry_count",
          "max_retries",
          "scheduled_at",
          "created_at",
          "updated_at"
        ],
        "title": "NotificationResponse"
      },
      "PaginatedResponse_NotificationResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[NotificationResponse]"
      },
...
      "PreferenceUpdate": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/ChannelEnum"
          },
          "opted_in": {
            "type": "boolean",
            "title": "Opted In"
          }
        },
        "type": "object",
        "required": [
          "channel",
          "opted_in"
        ],
        "title": "PreferenceUpdate"
      },
      "PriorityEnum": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "normal",
          "low"
        ],
        "title": "PriorityEnum"
      },
      "StatusEnum": {
        "type": "string",
        "enum": [
          "pending",
          "queued",
          "processing",
          "sent",
          "delivered",
          "failed"
        ],
        "title": "StatusEnum"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}